feat: Standardize sepeal and appeal data structures
The way sepeal and appeal data was being handled and stored was inconsistent. This change standardizes the field names for `assigned_to_id` and `recipient_id` to `assignedToId` and `recipientId` respectively, aligning with the `UserProfile` type and improving data consistency across the application.

It also ensures `created_at` is consistently handled.